// modular/packages/client/src/generated/routes.ts
// Generated by script/generate.ts. Do not edit.
export const routes = {
  "layout": {
    "method": "GET",
    "path": "/api/cybermastery/layout"
  },
  "save": {
    "method": "PUT",
    "path": "/api/cybermastery/layout"
  },
  "events": {
    "method": "GET",
    "path": "/api/cybermastery/events"
  }
} as const